requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6

# AI/ML
openai==1.55.3
sentence-transformers==3.3.1

# Document Processing
pypdf==5.1.0

# Data Validation
pydantic==2.10.3
